+++ richdoc/richdoc-cli/src/richdoc_cli/cli.py
@@ -18,9 +18,9 @@ from .commands.export import group as export_group
 from .commands.init_ import cmd as init_cmd
 from .commands.lint import cmd as lint_cmd
 from .commands.new_ import cmd as new_cmd
-from .commands.publish import group as publish_group
 from .commands.update import cmd as update_cmd
 from .output import json_error
+from .publish.confluence.cli import confluence_group
 
 
 @click.group(
@@ -42,7 +42,7 @@ main.add_command(update_cmd)
 main.add_command(lint_cmd)
 main.add_command(components_cmd)
 main.add_command(export_group)
-main.add_command(publish_group)
+main.add_command(confluence_group)
 
 
 def entrypoint() -> NoReturn: